parse transitive json file
